requirements.txt
python-telegram-bot
aiohttp
requests
pytgcalls
ffmpeg
yt-dlp
